Add files via upload
Mam, here is the link for the problem I had with Project 30. Please see if you can get the output in your system and help me out with the project. I had booked a project booster class for 3rd, so please help me with this as well as the other pending projects. Although this one is more important as the deadline for the rest has passed and my other teachers couldn't resolve my issues and it was always left pending.